move Error to Errors
- make close react to individual errors
- remove SecondColumn, we don't need it any longer
- add actions logger to show which error is being closed in storybook stories
- test for correct close parameter called
- update snaps
