+++ unlock-app/src/stories/interface/Error.stories.js
@@ -1,17 +1,19 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import { Errors } from '../../components/interface/Error'
-
-const close = () => {}
+import { action } from '@storybook/addon-actions'
+import { Errors } from '../../components/interface/Errors'
 
 const errors = ['error 1', 'error 2', 'error 3']
 
-storiesOf('Error', module)
+storiesOf('Errors', module)
   .add('Simple Error', () => {
     return (
-      <Errors close={close} error="We could not process that transaction." />
+      <Errors
+        close={action('close')}
+        errors={['We could not process that transaction.']}
+      />
     )
   })
   .add('Multiple errors', () => {
-    return <Errors errors={errors} close={close} />
+    return <Errors errors={errors} close={action('close')} />
   })
